Make DaemonCli really be reloaded and discarded along with CL
