+++ common/src/main/java/org/mvndaemon/mvnd/common/MavenDaemon.java
@@ -15,66 +15,10 @@
  */
 package org.mvndaemon.mvnd.common;
 
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.stream.Stream;
-
 public class MavenDaemon {
 
     public static void main(String[] args) throws Exception {
-
-        final Path mvndHome = Environment.MVND_HOME.asPath();
-        URL[] classpath = Stream.concat(
-                /* jars */
-                Stream.of("mvn/lib/ext", "mvn/lib", "mvn/boot")
-                        .map(mvndHome::resolve)
-                        .flatMap((Path p) -> {
-                            try {
-                                return Files.list(p);
-                            } catch (java.io.IOException e) {
-                                throw new RuntimeException("Could not list " + p, e);
-                            }
-                        })
-                        .filter(p -> {
-                            final String fileName = p.getFileName().toString();
-                            return fileName.endsWith(".jar") && !fileName.startsWith("mvnd-client-");
-                        })
-                        .filter(Files::isRegularFile),
-                /* resources */
-                Stream.of(mvndHome.resolve("mvn/conf"), mvndHome.resolve("mvn/conf/logging")))
-
-                .map(Path::normalize)
-                .map(Path::toUri)
-                .map(uri -> {
-                    try {
-                        return uri.toURL();
-                    } catch (MalformedURLException e) {
-                        throw new RuntimeException(e);
-                    }
-                })
-                .toArray(URL[]::new);
-        ClassLoader loader = new URLClassLoader(classpath, null) {
-            @Override
-            protected Class<?> findClass(String name) throws ClassNotFoundException {
-                try {
-                    return super.findClass(name);
-                } catch (ClassNotFoundException e) {
-                    return MavenDaemon.class.getClassLoader().loadClass(name);
-                }
-            }
-
-            @Override
-            public URL getResource(String name) {
-                URL url = super.getResource(name);
-                if (url == null) {
-                    url = MavenDaemon.class.getClassLoader().getResource(name);
-                }
-                return url;
-            }
-        };
+        ClassLoader loader = ClassLoaderHelper.createLoader(s -> true, null);
         Thread.currentThread().setContextClassLoader(loader);
         Class<?> clazz = loader.loadClass("org.mvndaemon.mvnd.daemon.Server");
         try (AutoCloseable server = (AutoCloseable) clazz.getConstructor().newInstance()) {
